requirements.txt
<<<<<<< HEAD
flask
gunicorn
PyPDF2
pdf2docx
python-docx
docx2pdf
Pillow
pandas
pdfplumber
reportlab

=======
flask
PyPDF2
pdf2docx
python-docx
docx2pdf
pandas
openpyxl
pillow
reportlab
pdfplumber
>>>>>>> 00799e05e3e02922ce1f2a9b1dd634d6a050d2bc
